Move app store init code
